Delete Opportunities column + Improving design of BoardColumnEditTitleMenu
