+++ front/src/modules/ui/board/components/BoardColumnEditTitleMenu.tsx
@@ -1,39 +1,58 @@
-import { ChangeEvent, useState } from 'react';
+import { ChangeEvent, useCallback, useState } from 'react';
 import styled from '@emotion/styled';
+import { useRecoilState } from 'recoil';
 
 import { StyledDropdownMenuItemsContainer } from '@/ui/dropdown/components/StyledDropdownMenuItemsContainer';
 import { StyledDropdownMenuSeparator } from '@/ui/dropdown/components/StyledDropdownMenuSeparator';
+import { IconTrash } from '@/ui/icon';
+import { MenuItem } from '@/ui/menu-item/components/MenuItem';
 import { MenuItemSelectColor } from '@/ui/menu-item/components/MenuItemSelectColor';
 import { ThemeColor } from '@/ui/theme/constants/colors';
 import { textInputStyle } from '@/ui/theme/constants/effects';
 import { debounce } from '~/utils/debounce';
+
+import { boardColumnsState } from '../states/boardColumnsState';
 
 const StyledEditTitleContainer = styled.div`
   --vertical-padding: ${({ theme }) => theme.spacing(1)};
 
   align-items: center;
 
+  border: ${({ theme }) => theme.border.radius.md};
+  /* border-color: pink; */
+  /* border-style: solid; */
+  /* border-width: 1px; */
   display: flex;
   flex-direction: row;
   height: calc(36px - 2 * var(--vertical-padding));
+
   padding: var(--vertical-padding) 0;
 
   width: calc(100%);
 `;
 
 const StyledEditModeInput = styled.input`
-  font-size: ${({ theme }) => theme.font.size.sm};
+  /* border: 1px solid white; */
+  background-color: pink;
 
   ${textInputStyle}
 
+  border-color: ${({ theme }) => theme.border.color.strong};
+  border-radius: ${({ theme }) => theme.border.radius.sm};
+  border-style: solid;
+  border-width: 1px;
+  font-size: ${({ theme }) => theme.font.size.sm};
+  height: 100%;
   width: 100%;
 `;
 
 export type BoardColumnEditTitleMenuProps = {
   onClose: () => void;
+  onDelete?: (id: string) => void;
   title: string;
   onTitleEdit: (title: string, color: string) => void;
   color: ThemeColor;
+  stageId: string;
 };
 
 type ColumnColorOption = {
@@ -56,6 +75,8 @@ export const COLUMN_COLOR_OPTIONS: ColumnColorOption[] = [
 
 export const BoardColumnEditTitleMenu = ({
   onClose,
+  onDelete,
+  stageId,
   onTitleEdit,
   title,
   color,
@@ -69,6 +90,16 @@ export const BoardColumnEditTitleMenu = ({
     setInternalValue(event.target.value);
     debouncedOnUpdateTitle(event.target.value);
   };
+  const [, setBoardColumns] = useRecoilState(boardColumnsState);
+
+  const handleDelete = useCallback(() => {
+    setBoardColumns((previousBoardColumns) =>
+      previousBoardColumns.filter((column) => column.id !== stageId),
+    );
+    onDelete?.(stageId);
+    onClose();
+  }, [onClose, onDelete, setBoardColumns, stageId]);
+
   return (
     <StyledDropdownMenuItemsContainer>
       <StyledEditTitleContainer>
@@ -92,6 +123,13 @@ export const BoardColumnEditTitleMenu = ({
           text={colorOption.name}
         />
       ))}
+      <StyledDropdownMenuSeparator />
+      <MenuItem
+        onClick={handleDelete}
+        LeftIcon={IconTrash}
+        text="Delete"
+        accent="danger"
+      />
     </StyledDropdownMenuItemsContainer>
   );
 };
