+++ front/src/modules/ui/table/editable-cell/type/components/GenericEditableDoubleTextChipCellDisplayMode.tsx
@@ -1,6 +1,5 @@
 import { useRecoilState } from 'recoil';
 
-
 import { DoubleTextChipDisplay } from '@/ui/content-display/components/DoubleTextChipDisplay';
 import { ViewFieldDoubleTextChipMetadata } from '@/ui/editable-field/types/ViewField';
 import { useCurrentRowEntityId } from '@/ui/table/hooks/useCurrentEntityId';
